Update extension manifest (#65)
Updated manifest to use the correct information and format. Have to
shorten the Description because I was getting an error complaining that
it's too long when I tried to publish it via the CI.

Also turning off completions since that is not currently supported, but
from the output channel I can see Agent kept sending us completions
related requests.