Workaround for Agent bug after calling Initialize() (#93)
CLOSE
https://linear.app/sourcegraph/issue/CODY-3637/visual-studio-loading-cody-chat-is-not-hiding

### The Agent initialization code has a bug, and this is workaround for
it:

When IAgentService.Initialize() is called, Agent doesn't wait to finish
executing Initialize() but calls `webview/registerWebviewViewProvider`
in the meantime ,without properly initializing
`webviewProtocolDelegate`.

Because `webview/registerWebviewViewProvider` is called on the client
side, client decides to call `webview/resolveWebviewView` but it could
fail when `webviewProtocolDelegate` is null. And
`webviewProtocolDelegate` is set during executing `Initialize()` (when
`registerNativeWebviewHandlers()` is called) on agent-side, and it will
be null, unless `IAgentService.Initialize()` will finish executing on
agent-side.