Dev: Allow specifying Cody branch or tag to build agent
This change allows specifying the Cody branch or tag to use when building the Cody agent, instead of always using the latest stable release tag.

The `cody-branch` argument has been added to the `BuildCodyAgent` task, which defaults to the latest stable release tag (`vscode-v1.38.1`) if not provided.

This provides more flexibility in building the agent, allowing developers to test against specific Cody branches or tags as needed.

CLOSE: https://linear.app/sourcegraph/issue/CODY-4069/run-build-off-cody-main-nightly-for-visual-studio