Adding tracing (#121)
The change aims to add detailed tracing to the extension's code.
Currently, we use logging to the VSs output window, but this has
limitations and should be used for conveying basic information.
Additionally, the output window is visible to users of the extension, so
it should not contain too many entries and details. The place where
details should be logged is tracing.

- Capability for dynamically enabling and disabling tracing
- When tracing is disabled, it does not impact performance
- Ability to filter entries additionally
- File logging (possibility to easy add custom targets where entries are
logged)
- Ability to view state of the extension in real time using external
tools

![Zrzut ekranu 2024-10-22
142258](https://github.com/user-attachments/assets/404b3efe-f496-42bc-8fcf-0c24309170e0)

![Zrzut ekranu 2024-10-22
142329](https://github.com/user-attachments/assets/6689bc55-4487-4b2c-8564-0c14e7eebb70)

---------

Co-authored-by: Tomasz Gołębiowski <tgolebiowski@virtuslab.com>