Sentry time bomb (#229)
Disable error logging to Sentry after 180 days from build date. The
change is intended to limit the number of events sent. We are mainly
interested in the latest versions of the extension so older versions
should be out of the scope after six months.
## Test plan
Check if after 6 months the extension version logs any errors to Sentry
😉
<!-- REQUIRED; info at
https://docs-legacy.sourcegraph.com/dev/background-information/testing_principles
-->

---------

Co-authored-by: Tomasz Gołębiowski <tgolebiowski@virtuslab.com>