Prevent webview to overlap other controls (#286)
The chat uses WebView2, which has always painted itself as the top
window enclosing everything underneath. This was a known issue that,
after years, was finally resolved by Microsoft
(https://github.com/MicrosoftEdge/WebView2Feedback/issues/286). This PR
changes WebView2 controls to the latest version that respects z-order in
WPF.
## Test plan
Try opening another window above the chat window or docking it to the
side of the ide and expanding it over the chat. In either case, the
window above the chat should not be overlapped by the chat.
<!-- REQUIRED; info at
https://docs-legacy.sourcegraph.com/dev/background-information/testing_principles
-->

---------

Co-authored-by: Tomasz Gołębiowski <tgolebiowski@virtuslab.com>