Use of the shutdown state in Cody logic (#306)
Add support for Visual Studio shutdown state in the extension. So far,
Cody has not used the information that the IDE is about to shut down and
tried to continue working. This led to many resource dispose exceptions
during app exit. This PR adds support for shutdown state and reduces
number of errors logged to Sentry.
## Test plan
N/A
<!-- REQUIRED; info at
https://docs-legacy.sourcegraph.com/dev/background-information/testing_principles
-->

Co-authored-by: Tomasz Gołębiowski <tgolebiowski@virtuslab.com>