WebView2 library downgrade (#329)
After updating `Microsoft.Web.WebView2` to version `1.0.3296.44` some
users have started reporting problems with the chat rendering (an empty
Cody window).

Looks like last update to `1.0.3351.48` doesn't resolve the issue also,
so downgrading to `1.0.2592.51` which is the last good working version.
Unfortunately it does not have fix for
https://github.com/MicrosoftEdge/WebView2Feedback/issues/286 (WebView2
is always on the top of the other controls).

Users are reporting a blank Cody window, but looks like internally that
the page is loaded, and when interacting with the control the cursor is
changing and reacting to different page areas (e. g. buttons).

More context:
https://sourcegraph.slack.com/archives/C073U3KE5DK/p1753121170575139

Probably related:
1. https://github.com/MicrosoftEdge/WebView2Feedback/issues/5189
2. https://github.com/MicrosoftEdge/WebView2Feedback/issues/5281