Smart apply (#326)
This PR adds the possibility of dynamically insert code from Cody chat
into files in solution. This can be accomplished by using Smart apply,
inserting code at cursor or saving code to a new file. Each of these
options is available under the code snippet in the chat window.

## Test plan
1. Use chat so that Cody suggests a piece of code
2. Under the code snippet, use the buttons:
- Apply - code should be smartly inserted into an already existing file
   -  Insert - code should be inserted in place of the cursor
- Save - after selecting the location and file name, it will be created
with the suggested code

<!-- REQUIRED; info at
https://docs-legacy.sourcegraph.com/dev/background-information/testing_principles
-->

---------

Co-authored-by: Tomasz Gołębiowski <tgolebiowski@virtuslab.com>