Support for notifications (#355)
This PR adds support for notifications from the agent. They can appear
as text in the status bar or as a window popping up in the lower right
corner of the main window. The window version appears when the
notification contains actions. In this case, they appear as buttons in
the window.
## Test plan
1. Log out of Cody
2. Log in again
3. A notification should appear in the status bar: "Signed in to `url`"
<!-- REQUIRED; info at
https://docs-legacy.sourcegraph.com/dev/background-information/testing_principles
-->

---------

Co-authored-by: Tomasz Gołębiowski <tgolebiowski@virtuslab.com>