Cody's status on statusbar (#371)
This PR adds a status icon to the status bar. The icon informs the user
about Cody's status (ready, working, error, etc.). The status is fully
controlled by the agent. In addition to the icon, it is possible to add
a caption and a longer description in a tooltip. The user can click on
the icon at any time to open a chat window.
## Test plan
Open VS and observe how the status icon in the status bar changes during
the initialization steps.
You can also log out of Cody to see different status icons.
<!-- REQUIRED; info at
https://docs-legacy.sourcegraph.com/dev/background-information/testing_principles
-->

---------

Co-authored-by: Tomasz Gołębiowski <tgolebiowski@virtuslab.com>