Install .net 6 as a step of PrivateGalleryCreator (#383)
Version 6 of .NET has been removed from the default image used for
publishing Cody. This version of the runtime is used by the
PrivateGalleryCreator feed generation tool. Until the author updates the
tool to the latest runtime version, we will install version 6. Cody does
NOT use this version of .NET.
## Test plan
N/A
<!-- REQUIRED; info at
https://docs-legacy.sourcegraph.com/dev/background-information/testing_principles
-->

Co-authored-by: Tomasz Gołębiowski <tgolebiowski@virtuslab.com>