Restoring the status icon in VS 2026 (#385)
The new version of Visual Studio changed the way the main window loads,
which resulted in the status icon not being visible in the status bar.
This PR fixes the situation so that the icon is visible in both versions
2022 and 2026.
## Test plan
Launch Visual Studio 2026 and check if the status icon appears in the
status bar.
<!-- REQUIRED; info at
https://docs-legacy.sourcegraph.com/dev/background-information/testing_principles
-->

---------

Co-authored-by: Tomasz Gołębiowski <tgolebiowski@virtuslab.com>